latest changes

+++ src/infrastructure/database/entities/AiTransactionsLogs.ts
@@ -13,7 +13,7 @@ import { isTestingEnv } from 'src/infrastructure/helpers/helpers';
 @Index('PK_AITransactionsLogs', ['id'], { unique: true })
 @Entity('AITransactionsLogs', { schema: 'user' })
 export class AiTransactionsLogs {
-  @PrimaryGeneratedColumn({ type: 'bigint', name: 'ID' })
+  @PrimaryGeneratedColumn({ type: 'int', name: 'ID' })
   id: string;
 
   @Column(isTestingEnv ? 'text' : 'uniqueidentifier', { name: 'ServiceInstanceID' })
